// client/app/shared/models/track.model.ts
export class Track {
  _id?: string;
  customer_name?: string;
  phone_no?: string;
  customer_address?: string;
  email?: string;
  date_registion?: Date;
  serial_no?: string;
  machine_type?: string;
  device_details?: string;
  fault_reported?: string;
  other_details?: string;
  registered_by?: string;
  engineer_in_charge?: string;
  date_for_collection?: string;
}